GTA04A3: Allow "Cannot write control '2:0:0:Codec Operation Mode:0' : Device or resource busy" output from alsactl
Previously we'd kill all soundcard users if there was any output at
all from alsactl.  Killing soundcard users is possibly a bad idea in
any case, but it's especially so when the output is actually benign.
For example, it means that the Media Player dies when headphones are
inserted or removed.

Regarding the "Codec Operation Mode" setting, Neil Brown recently
wrote:

> The twl4030 audio codec has 2 operation modes.
> In one the PCM (digital) connections to the bluetooth and Modem are
> disabled.  In the other they are enabled.
> The first is sometimes described as "multimedia player" mode while the
> seconds is "voice" mode.  I think you a limited to lower sample rates in
> voice mode.
>
> You cannot switch modes while the device is active.  i.e. nothing can have
> the device open to play or record.
>
> This error just tells you that it cannot switch mode.  It shouldn't stop
> other settings from being applied.
>
> On the A3 you should never need to switch modes (unless you are playing with
> using the PCM path to get sound to/from the bluetooth).
> On the A4 you need to either leave it in the second mode (which probably
> wastes power, and probably forces low data rates) or switch on the second mode
> at some safe point - maybe between stopping the ring and answering the call,
> or maybe before making the phone ring.

On the A3 board, all the possible ALSA states (i.e. all
except *hw.state) have 'Codec Operation Mode' = 'Option 1 (audio)', so
there are no transitions that actually switch the mode, and it is
not a problem for that setting to fail.  Therefore we should allow
that output and _not_ go killing soundcard users.

This change allows media playback to continue when headphones are
inserted or removed.  It also seems to make Voice Notes work more
reliably, and I suspect it might help with call audio too.

(Unfortunately this change does nothing yet for the A4 board.  On the
A4 things are more complex, as Neil Brown says above, because there
there are some ALSA state transitions that _do_ need to change the
codec operation mode.  So for A4 more thought and work are needed.)